refactor: move ownerPropertyScroll to profileID-page
